LayoutTests/http/tests/site-isolation/inspector/css/css-stylesheet-commands-frame-target.html

<script src="../../../inspector/resources/inspector-test.js"></script>
<script src="../debugger/resources/site-isolation-debugger-test-utilities.js"></script>
<script>
    function test() {
        let suite = InspectorTest.createAsyncSuite("SiteIsolation.CSS.StyleSheetCommands.FrameTarget");

        let childFrameTarget;
        let stylesheetId;

        suite.addTestCase({
            name: "SiteIsolation.CSS.StyleSheetCommands.FrameTarget.Setup",
            description: "Find the child frame target.",
            async test() {
                let frameTargets = WI.targets.filter((t) => t.type === WI.TargetType.Frame);
                let urlTargetMap = await fetchDocumentURLsForTargets(frameTargets);
                let mainFrameTarget = urlTargetMap.get(WI.networkManager.mainFrame.url);
                childFrameTarget = frameTargets.find((t) => t !== mainFrameTarget);
                InspectorTest.expectNotNull(childFrameTarget, "Should find a child frame target.");
                InspectorTest.expectNotNull(childFrameTarget.CSSAgent, "Child frame target should have a CSSAgent.");
            }
        });

        suite.runTestCasesAndFinish();
    }
</script>

<body onload="runTest()">
    <p>Test that CSS stylesheet commands work correctly on frame targets under site isolation.</p>
    <iframe src="http://localhost:8000/site-isolation/inspector/css/resources/frame-with-stylesheet.html"></iframe>
</body>
