LayoutTests/http/tests/site-isolation/inspector/css/css-stylesheet-commands-frame-target-expected.txt

Test that CSS stylesheet commands work correctly on frame targets under site isolation.



== Running test suite: SiteIsolation.CSS.StyleSheetCommands.FrameTarget
-- Running test case: SiteIsolation.CSS.StyleSheetCommands.FrameTarget.Setup
PASS: Should find a child frame target.
PASS: Child frame target should have a CSSAgent.
